% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_krl_station}
\alias{find_krl_station}
\title{Find KRL station name}
\usage{
find_krl_station(sts_name, type = "name", max = 2, ...)
}
\arguments{
\item{type}{type of station name input (\code{name}, \code{code} or \code{id})}

\item{...}{other \code{agrep} additional parameters}

\item{station_name}{KRL station name or keyword}
}
\description{
Find KRL station name using \code{agrep} R base function.
}
\examples{
\dontrun{
find_krl_station("stasiun sudirman")
}
}
