% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_krl_price}
\alias{get_krl_price}
\title{Get KRL price}
\usage{
get_krl_price(sts_origin, sts_dest, type = "name")
}
\arguments{
\item{sts_origin}{origin KRL station name}

\item{sts_dest}{destination KRL station name}

\item{type}{type of station name input (\code{name}, \code{code} or \code{id})}
}
\description{
Get KRL price based on \code{krl_price} dataset. Use \code{find_krl_station} to find the
KRL station exact name, code or ID as the input of this function.
}
\examples{
\dontrun{
get_krl_price("Ancol", "Sudirman")
}
}
