% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{id_map}
\alias{id_map}
\title{Get spatial data frame of Indonesia or Jakarta map}
\usage{
id_map(region = "jakarta", level = "default")
}
\arguments{
\item{region}{\code{jakarta} or \code{indonesia}}

\item{level}{\code{kelurahan}, \code{rw} and \code{kecamatan} for \code{jakarta} region (\code{default} is \code{kelurahan}).
\code{kota} or \code{provinsi} for \code{indonesia} region (\code{default} is \code{provinsi}).}
}
\description{
Get spatial data frame of Indonesia or Jakarta map
}
\examples{
id_map("indonesia", "kota")
id_map("jakarta", "rw")
}
