% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{id_to_en}
\alias{id_to_en}
\alias{id_to_en}
\alias{en_to_id}
\title{Translate day/month name (Indonesia <-> English)}
\usage{
id_to_en(x)

en_to_id(x)
}
\arguments{
\item{x}{character input}
}
\description{
Translate day/month name (Indonesia <-> English)
}
\examples{
x <- "1 Januari 2020"
y <- id_to_en(x)
y
en_to_id(y)
}
