% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kata_setop}
\alias{kata_setop}
\title{Indonesian stop words data}
\format{A tibble containing lists of Indonesian stop words}
\usage{
kata_setop
}
\description{
Lists of Indonesian stop words
}
\keyword{datasets}
