% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kota}
\alias{kota}
\title{Indonesia city and regency data}
\format{A data frame with two variabels: \code{id}, \code{kota}}
\source{
\href{https://github.com/edwardsamuel/Wilayah-Administratif-Indonesia/blob/master/csv/regencies.csv}{https://github.com/edwardsamuel/Wilayah-Administratif-Indonesia/blob/master/csv/regencies.csv}
}
\usage{
kota
}
\description{
Lists of Indonesia cities and regencies with the ID.
}
\references{
\href{http://mfdonline.bps.go.id/}{http://mfdonline.bps.go.id/}
}
\keyword{datasets}
