% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{krl_route}
\alias{krl_route}
\title{KRL route data}
\format{A tibble containing lists of KRL route}
\usage{
krl_route
}
\description{
Lists of KRL route with geometry (as per 25 June 2018)
}
\keyword{datasets}
