% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_map.R
\name{map_download}
\alias{map_download}
\alias{map_convert}
\title{Get Indonesia map from GADM}
\usage{
map_download(folder = paste0(system.file("data", package = "nusantr"),
  "/raw"), filename = "file.zip",
  url = "http://biogeo.ucdavis.edu/data/gadm2.8/shp/IDN_adm_shp.zip")

map_convert(folder = paste0(system.file("data", package = "nusantr"),
  "/raw"), filename = "file.zip", output = system.file("data/spatial",
  package = "nusantr"))
}
\arguments{
\item{folder}{download folder location}

\item{filename}{name of the file (zip)}

\item{url}{url location of the file}

\item{output}{output location}
}
\description{
Download and get Indonesia map from
\href{[http://biogeo.ucdavis.edu/data/gadm2.8/shp/IDN_adm_shp.zip]}{http://biogeo.ucdavis.edu/data/gadm2.8/shp/IDN_adm_shp.zip}
and convert them to spatial data frame (sf).
}
