% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nama}
\alias{nama}
\title{Popular name in Jakarta}
\format{A tibble containing popular name in Jakarta with 9,083 rows and 3 columns: \code{nama}, \code{jenis_kelamin} and \code{n}.
\code{nama} is the name, \code{jenis_kelamin} is the gender ("L" for male, "P" for female) and \code{n} is the number of occurence.}
\source{
\href{https://pilkada2017.kpu.go.id}{https://pilkada2017.kpu.go.id}
}
\usage{
nama
}
\description{
Lists of popular name in Jakarta with number of occurence more than 50.
The name only contains voters of Pilkada DKI Jakarta phase 2 year 2017.
}
\keyword{datasets}
