% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nik_to_all}
\alias{nik_to_all}
\alias{nik_to_all}
\alias{nik_to_gender}
\alias{nik_to_bd}
\alias{nik_to_city}
\alias{nik_to_prov}
\title{Parse gender, birthdate, geolevel (city/regency/province) from NIK}
\usage{
nik_to_all(nik, select = c("gender", "bd", "city", "prov"),
  typec = "default", typep = "default")

nik_to_gender(nik, return = "default")

nik_to_bd(nik)

nik_to_city(nik, type = "default")

nik_to_prov(nik, type = "default")
}
\arguments{
\item{nik}{NIK (KTP ID)}

\item{select}{select (\code{gender}, \code{bd}, \code{city} or \code{prov})}

\item{typec}{return format for city (\code{default}, \code{lc} or \code{short})}

\item{typep}{return format for province (\code{default}, \code{lc}, \code{short} or \code{abb})}

\item{return}{\code{nik_to_gender} return type (\code{default}, \code{short} or \code{abb})}

\item{type}{return format (\code{default}, \code{lc}, \code{short} or \code{abb})}
}
\value{
\itemize{
\item \code{nik_to_gender} returns a factor vector.
\item \code{nik_to_bd} returns a Date vector.
\item \code{nik_to_city} and \code{nik_to_prov} return a character vector.
\item \code{nik_to_all} returns a list of those four functions output.
}
}
\description{
NIK (\emph{Nomor Induk Kependudukan}) is a unique ID for Indonesia citizens
(\href{[https://en.wikipedia.org/wiki/Indonesian_identity_card]}{https://en.wikipedia.org/wiki/Indonesian_identity_card}).
NIK contains geographic location where the ID card is built, gender and birthdate of the holder.
}
\examples{
nik <- "1671065412990099"
nik_to_bd(nik)
nik_to_city(nik)
nik_to_prov(nik)
nik_to_all(nik)
nik_to_all(nik, select=c("gender","bd"))

# using nik_to_all
df <- data.frame(nik = rep("1671065412990099"), 5)
\dontrun{
df \%>\%
  mutate(out = map(nik, nik_to_all),
         out = map(out, data.frame)) \%>\%
  unnest()

# not using nik_to_all
df \%>\%
  mutate(bd = nik_to_bd(nik),
         city = nik_to_bd(city),
         prov = nik_to_bd(prov))
}
}
