% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nik_to_bd}
\alias{nik_to_bd}
\title{Get birth date based on NIK}
\usage{
nik_to_bd(nik)
}
\arguments{
\item{nik}{NIK (KTP ID)}
}
\description{
NIK (\emph{Nomor Induk Kependudukan}) is a unique ID for Indonesia citizens
(\href{[https://en.wikipedia.org/wiki/Indonesian_identity_card]}{https://en.wikipedia.org/wiki/Indonesian_identity_card}).
NIK contains geographic location where the ID card is built and the birth date of the citizen.
}
\examples{
nik_to_bd(1671065412990099)
#> [1] "1999-12-14"
}
