% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nik_to_city}
\alias{nik_to_city}
\alias{nik_to_prov}
\title{Get city/regency/province based on NIK}
\usage{
nik_to_city(nik)

nik_to_prov(nik)
}
\arguments{
\item{nik}{NIK (KTP ID)}
}
\description{
NIK (\emph{Nomor Induk Kependudukan}) is a unique ID for Indonesia citizens
(\href{[https://en.wikipedia.org/wiki/Indonesian_identity_card]}{https://en.wikipedia.org/wiki/Indonesian_identity_card}).
NIK contains geographic location where the ID card is built and the birth date of the citizen.
}
\examples{
nik_to_city(1671065412990099)
## [1] "KOTA PALEMBANG"
nik_to_prov(1671065412990099)
## [1] "SUMATERA SELATAN"
}
