% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinsi}
\alias{provinsi}
\title{Indonesia province data}
\format{A data frame with two variabels: \code{id}, \code{provinsi}}
\source{
\href{https://github.com/edwardsamuel/Wilayah-Administratif-Indonesia/blob/master/csv/provinces.csv}{https://github.com/edwardsamuel/Wilayah-Administratif-Indonesia/blob/master/csv/provinces.csv}
}
\usage{
provinsi
}
\description{
Lists of Indonesia provinces with the ID.
}
\references{
\href{http://mfdonline.bps.go.id/}{http://mfdonline.bps.go.id/}
}
\keyword{datasets}
