% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_spatial}
\alias{theme_spatial}
\title{Clean ggplot2 theme for geom_sf spatial visualization}
\usage{
theme_spatial(legend.inside = TRUE, base_family = "Roboto Condensed")
}
\arguments{
\item{legend.inside}{\code{TRUE}}

\item{base_family}{\code{Roboto Condensed} font}
}
\description{
Clean ggplot2 theme for geom_sf spatial visualization
}
