% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nik_to_all}
\alias{nik_to_all}
\alias{nik_to_all}
\alias{nik_to_bd}
\alias{nik_to_city}
\alias{nik_to_prov}
\title{Parse birthdate, geolevel (city/regency/province) from NIK}
\usage{
nik_to_all(nik)

nik_to_bd(nik)

nik_to_city(nik)

nik_to_prov(nik)
}
\arguments{
\item{nik}{NIK (KTP ID)}
}
\value{
\code{nik_to_bd} returns a Date vector.
\code{nik_to_city} and \code{nik_to_prov} return a character vector.
\code{nik_to_all} returns a list of those three functions output.
}
\description{
NIK (\emph{Nomor Induk Kependudukan}) is a unique ID for Indonesia citizens
(\href{[https://en.wikipedia.org/wiki/Indonesian_identity_card]}{https://en.wikipedia.org/wiki/Indonesian_identity_card}).
NIK contains geographic location where the ID card is built and the birthdate of the holder.
}
\examples{
nik <- "1671065412990099"
nik_to_bd(nik)
nik_to_city(nik)
nik_to_prov(nik)
nik_to_all(nik)

# using nik_to_all
df <- data.frame(nik = rep("1671065412990099"), 5)
df \%>\%
  mutate(out = map(nik, nik_to_all),
         out = map(out, data.frame)) \%>\%
  unnest()

# not using nik_to_all
df \%>\%
  mutate(bd = nik_to_bd(nik),
         city = nik_to_bd(city),
         prov = nik_to_bd(prov))
}
