% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transjakarta}
\alias{transjakarta}
\title{Transjakarta halte data}
\format{A tibble containing lists of Transjakarta halte}
\source{
\href{https://www.trafi.com/id/jakarta/transjakarta}{https://www.trafi.com/id/jakarta/transjakarta}
}
\usage{
transjakarta
}
\description{
Lists of Transjakarta halte with latitude and longitude (as per 25 June 2018)
}
\keyword{datasets}
